introduce sequence
